% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_light_validate.R
\name{light_validate}
\alias{light_validate}
\title{Function for validating CBAS Light (and secure) credential by checkig against stored procedure}
\usage{
light_validate(conn, group, Username, Password)
}
\arguments{
\item{conn}{a preconfigured connection object object using DBI and ODBC connection}

\item{group}{A CBAS Webapp group with permission to view the app}

\item{Username}{Usename}

\item{Password}{Password object passed in}
}
\value{
the result of a stored procedure validation call
}
\description{
Function for validating CBAS Light (and secure) credential by checkig against stored procedure
}
\examples{
# not run
#you will need to add some credential here you know should work 
# light_validate(conn = conn, group = "CBAS-WA-TEMPLATE",
#               Username = "smithj", Password = "<mypass>")
}
