% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_webapp_config.R
\name{webapp_config}
\alias{webapp_config}
\title{Function for creating all of the web_app connection component parts}
\usage{
webapp_config(
  uid = "webapptemplate",
  pwd = "UmmiesIdeK8#",
  group = "CBAS-WA-TEMPLATE",
  server = "CBAS-PDDB-06",
  db = "YOUR-DB-WKG"
)
}
\arguments{
\item{uid}{The webapp user ID}

\item{pwd}{The webapp password}

\item{group}{The group the webapp belongs too}

\item{server}{The server to connect to}

\item{db}{The database to query}
}
\value{
a list of webapp compnent parts
}
\description{
Users will need to to define the uid etc in the function in order to use and pass through.
}
\examples{
webapp_config()$server
}
