% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_webapp_conn.R
\name{webapp_conn}
\alias{webapp_conn}
\title{A function for establishing a connection to a database using a webapp credentials}
\usage{
webapp_conn(servername = "CBAS-PDDB-00", database = "CBAS-LIGHT", Uid, Pwd)
}
\arguments{
\item{servername}{servername you're connecting to}

\item{database}{The database you're attemtping to connect to}

\item{Uid}{Webapp user ID - supplied by CBAS}

\item{Pwd}{The webapp password - supplied by CBAS}
}
\value{
A connection object for querying a database
}
\description{
A function for establishing a connection to a database using a webapp credentials
}
\examples{
# conn <- webapp_conn(servername = "CBAS-PDDB-00",
}
